Uploading a new version
Updated version of the system to convert an Altium output file in .csv, into a file to be used in the NeoDen3v Pick and Place. The system now has a file search through a file manager window and tells you, in the main window, the name of the open file to avoid mistakes.